Major restructure of pyangbind directory for maintainability. 	* (M) migrated YANG type definitions into a separate lib/yangtypes 		  file, which can be imported by any output of pyangbind. 		  The file size for certain modules, and debugging of code in 		  this file was not optimal. 	* (M) restructure directories to have a common tests directory 		  rather than a lib/tests and tests/ separation. 	* (M) added ability to orchestrate tests across all modules for 		  commit hooks.
