--- lib/tests/02-static_ptr/xpathhelper.py
@@ -1,1 +1,0 @@
-../../xpathhelper.py
