+++ tests/decimal64/lib
@@ -1,0 +1,1 @@
+../../lib
