+++ tests/identityref/lib
@@ -1,0 +1,1 @@
+../../lib
