+++ tests/xpath/01-list_leaflist/lib
@@ -1,0 +1,1 @@
+../../../lib
