+++ tests/xpath/02-static_ptr/lib
@@ -1,0 +1,1 @@
+../../
