+++ tests/xpath/02-static_ptr/ptr-tc02.yang
