+++ tests/xpath/02-static_ptr/xpathhelper.py
